refactor: psalm burn pass 2a — nullsafety, truthiness, Geo dedupe; ledger 2,135→1,643 (#143)
- 208 PossiblyNullPropertyFetch sites converted to nullsafe access via
  psalm's exact column spans; 89 RiskyTruthyFalsyComparison sites
  rewritten as EXACT truthiness-preserving comparisons per operand type
  (int|null, string|null, float|null, arrays, false-unions).

- CRITICAL CATCH: the truthy transform mangled negated operands whose
  span included the '!' — including HasTeamFilters' tenancy scope, where
  the broken condition made the container-team fallback fire even when a
  session team exists (cross-tenant scope corruption in queue contexts).
  psalm's own paradox detection surfaced every damaged site; all
  repaired, and TeamScopeContainerFallbackTest now freezes the
  precedence (session team ALWAYS beats a stale container binding).
  Also repaired: report writers' summary sections (the broken condition
  silently skipped them) and NotificationController filters.

- App\Support\Geo: ONE haversine for the codebase — replaces five
  private copies (Route model, three monitoring jobs,
  RoutePlanningService), all-float so strict-operand findings die at
  the single source.

- App\Support\CurrentUser: get() narrows Auth::user() to ?User once;
  team() owns the null-guard so 41 per-site guard blocks collapse to
  one-liners that are clean under BOTH analyzers (phpstan requires the
  null handled; psalm's plugin knows currentTeam is never null).

- RoutePlanningService fully reshaped (78→0 findings): OSRM response
  ingestion guard-extracted with exact array shapes, malformed
  coordinates dropped, dead pathIntersectsGeofence/
  calculateAvoidancePoint deleted, unused machineId parameter removed.

- Integration family motif pass (37 foreach guards, 58 is_array wraps,
  data_get conversions); flat metric-map returns protected from the
  list-filter (ManufacturerMetricsShapeTest caught the one regression).

- psalter closure return types; 53 redundant casts removed; Team
  docblock aligned with the real schema (timezone/language/currency
  NOT NULL) and dead coalesces dropped.

Ratchet: PSALM_BASELINE_FILES_HIGH_WATER 168→163; occurrences
2,135→1,643. Both analyzers clean; 638 tests green (2 new).

+++ app/Jobs/ArchiveOldMetricsJob.php
@@ -37,8 +37,8 @@ class ArchiveOldMetricsJob implements ShouldQueue
 
     public function handle(): void
     {
-        $retentionDays = $this->retentionDays ?: (int) config('retention.metrics_days', 90);
-        $batchSize = $this->batchSize ?: (int) config('retention.archive_batch', 5000);
+        $retentionDays = $this->retentionDays ?: config('retention.metrics_days', 90);
+        $batchSize = $this->batchSize ?: config('retention.archive_batch', 5000);
         $cutoff = now()->subDays($retentionDays);
 
         Log::info('ArchiveOldMetricsJob: starting', [
